issue #1404: Add ios and mac os test scripts,update python and shell scr...

+++ tools/JenkinsScript/Mac/android/AndroidTest-release4_0-4_0_3.sh
@@ -6,9 +6,11 @@
 #Copy monkeyrunner python script to tools directory.
 cd ../..
 CUR=$(pwd)
-cp $CUR/Monkeyrunner_TestsCpp.py $ANDROID_HOME/tools
+cp $CUR/Monkeyrunner_TestCpp.py $ANDROID_HOME/tools
+cp $CUR/ReportManager.py $ANDROID_HOME/tools
 
 cd ../..
+PROJECT_HOME=$(pwd)
 cd samples/TestCpp/proj.android/bin
 
 #Copy test apk to tools directory.
@@ -17,22 +19,38 @@ cp $CUR/TestCpp-release-14.apk $ANDROID_HOME/tools
 cp $CUR/TestCpp-release-15.apk $ANDROID_HOME/tools
 
 #Enter tools directory.
-cd $ANDROID_HOME
+cd $ANDROID_HOME/tools
 
 #If monkeyrunner test failed,it automatically exit and make ERRORLEVEL nonzero.
 
 #Running monkeyrunner test(release,API level:14)
 mv TestCpp-release-14.apk TestCpp-release.apk
-monkeyrunner Monkeyrunner_TestsCpp.py
+monkeyrunner Monkeyrunner_TestCpp.py TestCpp-release.apk
+if [ $? != 0 ]; then
+    python ReportManager.py
+    exit 1
+fi
 rm TestCpp-release.apk
 
 #Running monkeyrunner test(release,API level:15)
 mv TestCpp-release-15.apk TestCpp-release.apk
-monkeyrunner Monkeyrunner_TestsCpp.py
+monkeyrunner Monkeyrunner_TestCpp.py TestCpp-release.apk
+if [ $? != 0 ]; then
+    python ReportManager.py
+    exit 1
+fi
 rm TestCpp-release.apk
 
-rm Monkeyrunner_TestsCpp.py
+rm Monkeyrunner_TestCpp.py
+rm ReportManager.py
 
 #Monkeyrunner success!
+echo Monkeyrunner Test Success!
+
+#Clean project files.
+cd $PROJECT_HOME
+
 git checkout -f
 git clean -df -x
+
+#End
